+++ src/lib/components/SceneProviders.svelte
@@ -24,9 +24,7 @@
 	import { provideFramelessComponents } from '$lib/hooks/useFramelessComponents.svelte'
 	import { provideResourceByName } from '$lib/hooks/useResourceByName.svelte'
 	import { provideWorld } from '$lib/ecs'
-	import { trait } from 'koota'
-
-	trait()
+	import { provide3DModels } from '$lib/hooks/use3DModels.svelte'
 
 	interface Props {
 		cameraPose?: CameraPose
@@ -52,6 +50,7 @@
 	provideResourceByName(() => partID.current)
 	provideFrames(() => partID.current)
 	provideGeometries(() => partID.current)
+	provide3DModels(() => partID.current)
 	providePointclouds(() => partID.current)
 	provideMotionClient(() => partID.current)
 	provideArmClient(() => partID.current)