+++ src/lib/ecs/useWorld.ts
@@ -4,9 +4,7 @@ import { getContext, setContext } from 'svelte'
 const key = Symbol('koota-context')
 
 export function provideWorld() {
-	console.log('hi')
 	const world = createWorld()
-	console.log(world)
 
 	setContext<World>(key, world)
 }