+++ src/lib/hooks/usePose.svelte.ts
@@ -9,8 +9,11 @@ import { useEnvironment } from './useEnvironment.svelte'
 import { observe } from '@threlte/core'
 import { untrack } from 'svelte'
 import { useFrames } from './useFrames.svelte'
+import { RefetchRates } from '$lib/components/RefreshRate.svelte'
+import { useLogs } from './useLogs.svelte'
 
 export const usePose = (name: () => string, parent: () => string | undefined) => {
+	const logs = useLogs()
 	const { refreshRates } = useMachineSettings()
 	const partID = usePartID()
 	const motionClient = useMotionClient()
@@ -32,15 +35,17 @@ export const usePose = (name: () => string, parent: () => string | undefined) =>
 	const options = $derived(
 		queryOptions({
 			enabled:
-				interval !== -1 &&
+				interval !== RefetchRates.OFF &&
 				client.current !== undefined &&
 				environment.current.viewerMode === 'monitor',
-			refetchInterval: interval === 0 ? false : interval,
-			queryKey: ['partID', partID.current, client.current?.name, 'getPose', name(), parent()],
+			refetchInterval: interval === RefetchRates.MANUAL ? false : interval,
+			queryKey: ['getPose', 'partID', partID.current, client.current?.name, name(), parent()],
 			queryFn: async () => {
 				if (!client.current) {
 					throw new Error('No client')
 				}
+
+				logs.add(`Fetching pose for ${name()}...`)
 
 				const resolvedParent = parentResource?.subtype === 'arm' ? `${parent()}_origin` : parent()
 				const pose = await client.current.getPose(name(), resolvedParent ?? 'world', [])
@@ -60,6 +65,12 @@ export const usePose = (name: () => string, parent: () => string | undefined) =>
 			}
 		}
 	)
+
+	$effect(() => {
+		if (query.current.error) {
+			logs.add(query.current.error.message, 'error')
+		}
+	})
 
 	return {
 		get current() {