+++ src/lib/hooks/useWeblabs.svelte.ts
@@ -3,6 +3,7 @@ import { SvelteSet } from 'svelte/reactivity'
 
 export const WEBLABS_EXPERIMENTS = {
 	MOTION_TOOLS_EDIT_FRAME: 'MOTION_TOOLS_EDIT_FRAME',
+	MOTION_TOOLS_RENDER_ARM_MODELS: 'MOTION_TOOLS_RENDER_ARM_MODELS',
 } as const
 
 export const WEBLABS_CONTEXT_KEY = Symbol('weblabs-context')
@@ -74,5 +75,12 @@ export const provideWeblabs = () => {
 }
 
 export const useWeblabs = () => {
-	return getContext<Context>(WEBLABS_CONTEXT_KEY)
+	const context = getContext<Context>(WEBLABS_CONTEXT_KEY)
+	if (!context) {
+		return {
+			load: () => {},
+			isActive: () => false,
+		}
+	}
+	return context
 }