+++ src/lib/lib.ts
@@ -6,6 +6,7 @@ export { default as AxesHelper } from './components/AxesHelper.svelte'
 export { BatchedArrow } from './three/BatchedArrow'
 export { CapsuleGeometry } from './three/CapsuleGeometry'
 export { OrientationVector } from './three/OrientationVector'
+export { WorldObject } from './WorldObject.svelte'
 
 // Functions
 export { parsePcdInWorker } from './loaders/pcd'